Create Get-SpongeText.ps1
Get-SpongeText.ps1 steps through all lowercase chars in a document and replaces the first instance with a capital letter with a 33% chance. It then steps through all uppercase chars and replaces the first instance with a lowercase letter with a 33% chance. 
Usage: Get-SpongeText.ps1 -folderPath <path to folder with .docs>